# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartLecture AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /quizzes:
    get:
      operationId: getQuizzes
      tags: [quizzes]
      summary: List all quizzes
      responses:
        "200":
          description: Quiz list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get a quiz with its questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDetail"
  /quizzes/{id}/attempt:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      summary: Submit answers and get score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, name, email, avatar, level, xp, streak]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        level:
          type: integer
        xp:
          type: integer
        streak:
          type: integer
        joinedAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        avatar:
          type: string
...
    Quiz:
      type: object
      required: [id, title, subject, subjectColor, questionCount, bestScore, attempts, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        subjectColor:
          type: string
        questionCount:
          type: integer
        bestScore:
          type: ["integer", "null"]
        attempts:
          type: integer
        createdAt:
          type: string
    QuizDetail:
      allOf:
        - $ref: "#/components/schemas/Quiz"
        - type: object
          required: [questions]
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
    QuizInput:
      type: object
      required: [title, subject]
      properties:
        title:
          type: string
        subject:
          type: string
        lectureId:
          type: integer
    Question:
      type: object
      required: [id, quizId, text, options, correctIndex]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        explanation:
          type: ["string", "null"]
    QuizAttemptInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: integer
    QuizResult:
      type: object
      required: [score, correctCount, totalCount, xpEarned, passed, questionResults]
      properties:
        score:
          type: integer
        correctCount:
          type: integer
        totalCount:
          type: integer
        xpEarned:
          type: integer
        passed:
          type: boolean
        questionResults:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResult"
    QuestionResult:
      type: object
      required: [questionId, correct, selectedIndex, correctIndex]
      properties:
        questionId:
          type: integer
        correct:
          type: boolean
        selectedIndex:
          type: integer
        correctIndex:
          type: integer
        explanation:
          type: ["string", "null"]
...