# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartLecture AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Quiz {
  id: number;
  title: string;
  subject: string;
  subjectColor: string;
  questionCount: number;
  /** @nullable */
  bestScore: number | null;
  attempts: number;
  createdAt: string;
}