# lib/api-zod/src/generated/types/quizInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartLecture AI API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizInput {
  title: string;
  subject: string;
  lectureId?: number;
}